+++ jdbc/src/main/java/org/apache/metamodel/jdbc/dialects/DefaultQueryRewriter.java
@@ -207,7 +207,7 @@ public class DefaultQueryRewriter extends AbstractQueryRewriter {
     }
 
     @Override
-    public boolean isFirstRowSupported() {
+    public boolean isFirstRowSupported(final Query query) {
         return false;
     }