+++ jdbc/src/main/java/org/apache/metamodel/jdbc/dialects/OracleQueryRewriter.java
@@ -30,7 +30,7 @@ public class OracleQueryRewriter extends OffsetFetchQueryRewriter {
     public static final int FIRST_FETCH_SUPPORTING_VERSION = 12;
 
     public OracleQueryRewriter(JdbcDataContext dataContext) {
-        super(dataContext, FIRST_FETCH_SUPPORTING_VERSION);
+        super(dataContext, FIRST_FETCH_SUPPORTING_VERSION, false);
     }
 
     @Override