+++ jdbc/src/test/java/org/apache/metamodel/jdbc/integrationtests/OracleTest.java
@@ -1,27 +1,23 @@
 /**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
+ * Licensed to the Apache Software Foundation (ASF) under one or more contributor license agreements.  See the NOTICE
+ * file distributed with this work for additional information regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the
+ * License.  You may obtain a copy of the License at
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations under the License.
  */
 package org.apache.metamodel.jdbc.integrationtests;
 
+import java.math.BigDecimal;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
@@ -32,6 +28,7 @@ import javax.swing.table.TableModel;
 import org.apache.metamodel.DataContext;
 import org.apache.metamodel.data.DataSet;
 import org.apache.metamodel.data.DataSetTableModel;
+import org.apache.metamodel.data.Row;
 import org.apache.metamodel.jdbc.JdbcDataContext;
 import org.apache.metamodel.jdbc.JdbcTestTemplates;
 import org.apache.metamodel.jdbc.dialects.IQueryRewriter;
@@ -49,25 +46,25 @@ import com.google.common.collect.Sets;
 /**
  * Test case that tests oracle interaction. An express edition of the oracle
  * database can be used to run these tests.
- * 
+ *
  * The test requires the "human resources" schema that is provided ass a sample
  * schema for Oracle default installations.
- * 
+ *
  * The script for installing it can be found in:
- * 
+ *
  * <pre>
  * $ORACLE_HOME / demo / schema / human_resources / hr_main.sql
  * </pre>
- * 
+ *
  * Install with something like:
- * 
+ *
  * <pre>
  * $ORACLE_HOME/bin/sqlplus -S &quot;/ as sysdba&quot; @hr_main.sql
  * </pre>
- * 
+ *
  * The JDBC driver is not available in the Maven repository so you will have to
  * download and attach it to the eclipse project yourself.
- * 
+ *
  * @see http://www.oracle.com/technology/products/bi/samples
  * @see http
  *      ://www.oracle.com/technology/software/products/database/xe/index.html
@@ -78,7 +75,7 @@ public class OracleTest extends AbstractJdbIntegrationTest {
     protected String getPropertyPrefix() {
         return "oracle";
     }
-    
+
     public void testGetQueryRewriter() throws Exception {
         if (!isConfigured()) {
             return;
@@ -95,12 +92,12 @@ public class OracleTest extends AbstractJdbIntegrationTest {
 
         JdbcTestTemplates.simpleCreateInsertUpdateAndDrop(getDataContext(), "metamodel_test_simple");
     }
-    
+
     public void testTimestampValueInsertSelect() throws Exception {
         if (!isConfigured()) {
             return;
         }
-        
+
         final Connection connection = getConnection();
         JdbcTestTemplates.timestampValueInsertSelect(connection, TimeUnit.MICROSECONDS, null);
     }
@@ -122,9 +119,11 @@ public class OracleTest extends AbstractJdbIntegrationTest {
             return;
         }
 
-        Schema schema = new JdbcDataContext(getConnection(), new TableType[] { TableType.TABLE }, null)
-                .getSchemaByName("SYS");
-        assertEquals(12, schema.getTableCount());
+        Schema schema =
+                new JdbcDataContext(getConnection(), new TableType[] { TableType.TABLE }, null).getSchemaByName("SYS");
+
+        // We cannot say anything about the correct count, just that there _must_ be tables in that schema
+        assertTrue(schema.getTableCount() > 0);
     }
 
     public void testGetSchemaNames() throws Exception {
@@ -168,8 +167,9 @@ public class OracleTest extends AbstractJdbIntegrationTest {
             String fkTableName = rs.getString(7);
             assertEquals("EMPLOYEES", fkTableName);
             String fkColumnName = rs.getString(8);
-            System.out.println("Found primary key relation: pkTableName=" + pkTableName + ",pkColumnName="
-                    + pkColumnName + ",fkTableName=" + fkTableName + ",fkColumnName=" + fkColumnName);
+            System.out.println(
+                    "Found primary key relation: pkTableName=" + pkTableName + ",pkColumnName=" + pkColumnName
+                            + ",fkTableName=" + fkTableName + ",fkColumnName=" + fkColumnName);
         }
         rs.close();
         assertEquals(3, count);
@@ -184,8 +184,9 @@ public class OracleTest extends AbstractJdbIntegrationTest {
             String fkTableName = rs.getString(7);
             assertEquals("DEPARTMENTS", fkTableName);
             String fkColumnName = rs.getString(8);
-            System.out.println("Found primary key relation: pkTableName=" + pkTableName + ",pkColumnName="
-                    + pkColumnName + ",fkTableName=" + fkTableName + ",fkColumnName=" + fkColumnName);
+            System.out.println(
+                    "Found primary key relation: pkTableName=" + pkTableName + ",pkColumnName=" + pkColumnName
+                            + ",fkTableName=" + fkTableName + ",fkColumnName=" + fkColumnName);
         }
         rs.close();
         assertEquals(2, count);
@@ -197,46 +198,45 @@ public class OracleTest extends AbstractJdbIntegrationTest {
         }
         Schema schema = getDataContext().getSchemaByName("HR");
         assertNotNull(schema);
-        assertEquals("{JdbcTable[name=COUNTRIES,type=TABLE,remarks=<null>],"
-                + "JdbcTable[name=DEPARTMENTS,type=TABLE,remarks=<null>]"
-                + ",JdbcTable[name=EMPLOYEES,type=TABLE,remarks=<null>]"
-                + ",JdbcTable[name=JOBS,type=TABLE,remarks=<null>]"
-                + ",JdbcTable[name=JOB_HISTORY,type=TABLE,remarks=<null>]"
-                + ",JdbcTable[name=LOCATIONS,type=TABLE,remarks=<null>]"
-                + ",JdbcTable[name=REGIONS,type=TABLE,remarks=<null>]"
-                + ",JdbcTable[name=EMP_DETAILS_VIEW,type=VIEW,remarks=<null>]}", Arrays.toString(schema.getTables().toArray()));
+
+        final List<String> expectedTableNames =
+                Arrays.asList("COUNTRIES", "DEPARTMENTS", "EMPLOYEES", "JOBS", "JOB_HISTORY", "LOCATIONS", "REGIONS",
+                        "EMP_DETAILS_VIEW");
+        final List<String> tableNames = schema.getTables().stream().map(Table::getName).collect(Collectors.toList());
+        assertTrue(tableNames.containsAll(expectedTableNames));
 
         Collection<Relationship> employeeRelationships = schema.getTableByName("EMPLOYEES").getRelationships();
 
-        Set<String> employeeRelStrings = employeeRelationships.stream().map(rel -> rel.toString()).collect(Collectors.toSet());
+        Set<String> employeeRelStrings =
+                employeeRelationships.stream().map(Object::toString).collect(Collectors.toSet());
 
         assertEquals(Sets.newHashSet(
-                "Relationship[primaryTable=EMPLOYEES,primaryColumns={EMPLOYEE_ID},foreignTable=DEPARTMENTS,foreignColumns={MANAGER_ID}]",
-                        "Relationship[primaryTable=DEPARTMENTS,primaryColumns={DEPARTMENT_ID},foreignTable=EMPLOYEES,foreignColumns={DEPARTMENT_ID}]",
-                        "Relationship[primaryTable=EMPLOYEES,primaryColumns={EMPLOYEE_ID},foreignTable=EMPLOYEES,foreignColumns={MANAGER_ID}]",
-                        "Relationship[primaryTable=JOBS,primaryColumns={JOB_ID},foreignTable=EMPLOYEES,foreignColumns={JOB_ID}]",
-                        "Relationship[primaryTable=EMPLOYEES,primaryColumns={EMPLOYEE_ID},foreignTable=JOB_HISTORY,foreignColumns={EMPLOYEE_ID}]"),
+                "Relationship[primaryTable=DEPARTMENTS,primaryColumns=[DEPARTMENT_ID],foreignTable=EMPLOYEES,foreignColumns=[DEPARTMENT_ID]]",
+                "Relationship[primaryTable=EMPLOYEES,primaryColumns=[EMPLOYEE_ID],foreignTable=DEPARTMENTS,foreignColumns=[MANAGER_ID]]",
+                "Relationship[primaryTable=EMPLOYEES,primaryColumns=[EMPLOYEE_ID],foreignTable=EMPLOYEES,foreignColumns=[MANAGER_ID]]",
+                "Relationship[primaryTable=JOBS,primaryColumns=[JOB_ID],foreignTable=EMPLOYEES,foreignColumns=[JOB_ID]]",
+                "Relationship[primaryTable=EMPLOYEES,primaryColumns=[EMPLOYEE_ID],foreignTable=JOB_HISTORY,foreignColumns=[EMPLOYEE_ID]]"),
                 employeeRelStrings);
 
         assertEquals(
-                "{JdbcColumn[name=EMPLOYEE_ID,columnNumber=0,type=DECIMAL,nullable=false,nativeType=NUMBER,columnSize=6],"
-                        + "JdbcColumn[name=FIRST_NAME,columnNumber=1,type=VARCHAR,nullable=true,nativeType=VARCHAR2,columnSize=20],"
-                        + "JdbcColumn[name=LAST_NAME,columnNumber=2,type=VARCHAR,nullable=false,nativeType=VARCHAR2,columnSize=25],"
-                        + "JdbcColumn[name=EMAIL,columnNumber=3,type=VARCHAR,nullable=false,nativeType=VARCHAR2,columnSize=25],"
-                        + "JdbcColumn[name=PHONE_NUMBER,columnNumber=4,type=VARCHAR,nullable=true,nativeType=VARCHAR2,columnSize=20],"
-                        + "JdbcColumn[name=HIRE_DATE,columnNumber=5,type=DATE,nullable=false,nativeType=DATE,columnSize=7],"
-                        + "JdbcColumn[name=JOB_ID,columnNumber=6,type=VARCHAR,nullable=false,nativeType=VARCHAR2,columnSize=10],"
-                        + "JdbcColumn[name=SALARY,columnNumber=7,type=DECIMAL,nullable=true,nativeType=NUMBER,columnSize=8],"
-                        + "JdbcColumn[name=COMMISSION_PCT,columnNumber=8,type=DECIMAL,nullable=true,nativeType=NUMBER,columnSize=2],"
-                        + "JdbcColumn[name=MANAGER_ID,columnNumber=9,type=DECIMAL,nullable=true,nativeType=NUMBER,columnSize=6],"
-                        + "JdbcColumn[name=DEPARTMENT_ID,columnNumber=10,type=DECIMAL,nullable=true,nativeType=NUMBER,columnSize=4]}",
+                "[Column[name=EMPLOYEE_ID,columnNumber=0,type=DECIMAL,nullable=false,nativeType=NUMBER,columnSize=6], "
+                        + "Column[name=FIRST_NAME,columnNumber=1,type=VARCHAR,nullable=true,nativeType=VARCHAR2,columnSize=20], "
+                        + "Column[name=LAST_NAME,columnNumber=2,type=VARCHAR,nullable=false,nativeType=VARCHAR2,columnSize=25], "
+                        + "Column[name=EMAIL,columnNumber=3,type=VARCHAR,nullable=false,nativeType=VARCHAR2,columnSize=25], "
+                        + "Column[name=PHONE_NUMBER,columnNumber=4,type=VARCHAR,nullable=true,nativeType=VARCHAR2,columnSize=20], "
+                        + "Column[name=HIRE_DATE,columnNumber=5,type=TIMESTAMP,nullable=false,nativeType=DATE,columnSize=7], "
+                        + "Column[name=JOB_ID,columnNumber=6,type=VARCHAR,nullable=false,nativeType=VARCHAR2,columnSize=10], "
+                        + "Column[name=SALARY,columnNumber=7,type=DECIMAL,nullable=true,nativeType=NUMBER,columnSize=8], "
+                        + "Column[name=COMMISSION_PCT,columnNumber=8,type=DECIMAL,nullable=true,nativeType=NUMBER,columnSize=2], "
+                        + "Column[name=MANAGER_ID,columnNumber=9,type=DECIMAL,nullable=true,nativeType=NUMBER,columnSize=6], "
+                        + "Column[name=DEPARTMENT_ID,columnNumber=10,type=DECIMAL,nullable=true,nativeType=NUMBER,columnSize=4]]",
                 Arrays.toString(schema.getTableByName("EMPLOYEES").getColumns().toArray()));
 
         assertEquals(
-                "{JdbcColumn[name=DEPARTMENT_ID,columnNumber=0,type=DECIMAL,nullable=false,nativeType=NUMBER,columnSize=4],"
-                        + "JdbcColumn[name=DEPARTMENT_NAME,columnNumber=1,type=VARCHAR,nullable=false,nativeType=VARCHAR2,columnSize=30],"
-                        + "JdbcColumn[name=MANAGER_ID,columnNumber=2,type=DECIMAL,nullable=true,nativeType=NUMBER,columnSize=6],"
-                        + "JdbcColumn[name=LOCATION_ID,columnNumber=3,type=DECIMAL,nullable=true,nativeType=NUMBER,columnSize=4]}",
+                "[Column[name=DEPARTMENT_ID,columnNumber=0,type=DECIMAL,nullable=false,nativeType=NUMBER,columnSize=4], "
+                        + "Column[name=DEPARTMENT_NAME,columnNumber=1,type=VARCHAR,nullable=false,nativeType=VARCHAR2,columnSize=30], "
+                        + "Column[name=MANAGER_ID,columnNumber=2,type=DECIMAL,nullable=true,nativeType=NUMBER,columnSize=6], "
+                        + "Column[name=LOCATION_ID,columnNumber=3,type=DECIMAL,nullable=true,nativeType=NUMBER,columnSize=4]]",
                 Arrays.toString(schema.getTableByName("DEPARTMENTS").getColumns().toArray()));
     }
 
@@ -248,12 +248,13 @@ public class OracleTest extends AbstractJdbIntegrationTest {
         Table employeeTable = schema.getTableByName("EMPLOYEES");
         Table departmentsTable = schema.getTableByName("DEPARTMENTS");
         Optional<Relationship> relationship = employeeTable.getRelationships(departmentsTable).stream()
-                .filter(rel-> rel.toString().equals("Relationship[primaryTable=EMPLOYEES,primaryColumns={EMPLOYEE_ID},foreignTable=DEPARTMENTS,foreignColumns={MANAGER_ID}]"))
+                .filter(rel -> rel.toString()
+                        .equals("Relationship[primaryTable=EMPLOYEES,primaryColumns=[EMPLOYEE_ID],foreignTable=DEPARTMENTS,foreignColumns=[MANAGER_ID]]"))
                 .findFirst();
         assertTrue(relationship.isPresent());
 
-        Query q = new Query().from(new FromItem(JoinType.INNER, relationship.get())).select(
-                employeeTable.getColumnByName("EMAIL"), departmentsTable.getColumnByName("DEPARTMENT_NAME"));
+        Query q = new Query().from(new FromItem(JoinType.INNER, relationship.get()))
+                .select(employeeTable.getColumnByName("EMAIL"), departmentsTable.getColumnByName("DEPARTMENT_NAME"));
         q.getSelectClause().getItem(0).setAlias("e-mail");
 
         assertEquals(
@@ -265,8 +266,40 @@ public class OracleTest extends AbstractJdbIntegrationTest {
         TableModel tableModel = new DataSetTableModel(data);
         assertEquals(2, tableModel.getColumnCount());
         assertEquals(11, tableModel.getRowCount());
-        assertEquals("JWHALEN", tableModel.getValueAt(0, 0).toString());
-        assertEquals("Administration", tableModel.getValueAt(0, 1).toString());
+
+        boolean found = false;
+        for (int i = 0; i < tableModel.getRowCount(); i++) {
+            if (tableModel.getValueAt(i, 0).toString().equals("JWHALEN")) {
+                assertEquals("Administration", tableModel.getValueAt(i, 1).toString());
+                found = true;
+                break;
+            }
+        }
+        assertTrue(found);
     }
 
+    public void testMaxAndOffset() throws Exception {
+        if (!isConfigured()) {
+            return;
+        }
+
+        final JdbcDataContext context = getDataContext();
+
+        final List<Row> onlyMaxRows =
+                context.query().from("HR", "EMPLOYEES").select("EMPLOYEE_ID").maxRows(10).execute().toRows();
+        assertEquals("Should limit size even without offset", 10, onlyMaxRows.size());
+
+        final List<Row> onlyOffset =
+                context.query().from("HR", "EMPLOYEES").select("EMPLOYEE_ID").orderBy("EMPLOYEE_ID").firstRow(5)
+                        .execute().toRows();
+        assertEquals("Should offset first row", new BigDecimal(104), onlyOffset.get(0).getValue(0));
+        assertEquals("Should not limit size beyond offset", 103, onlyOffset.size());
+
+        final List<Row> maxRowsAndOffset =
+                context.query().from("HR", "EMPLOYEES").select("EMPLOYEE_ID").maxRows(20).orderBy("EMPLOYEE_ID")
+                        .firstRow(20).execute().toRows();
+
+        assertEquals("Should offset first row", new BigDecimal(119), maxRowsAndOffset.get(0).getValue(0));
+        assertEquals("Should not limit size", 20, maxRowsAndOffset.size());
+    }
 }